change 05: slotmap generational handles, u32 → u64 FFI widening
Replace per-endpoint SlabSet + compose_handle/decompose_handle with global
slotmap registries using 64-bit generational handles (KeyData::as_ffi()).

Core (iroh-http-core):
- stream.rs: 7 new_key_type! declarations, 6 OnceLock<Mutex<SlotMap<K,…>>>
  global registries (reader, writer, trailer_tx, trailer_rx, fetch_cancel,
  session, request_head); ep_idx stored in Entry.ep_idx for sweep on
  unregister_endpoint(); all public handle types u32 → u64
- endpoint.rs: remove slabs: Arc<SlabSet> field; register_endpoint is no-op
- client.rs: alloc_fetch_token/cancel_in_flight/get_fetch_cancel_notify/
  remove_fetch_token delegated to stream module; fetch_token: Option<u64>
- server.rs: respond(req_handle: u64); local allocate_req_handle removed
  (now in stream.rs); on_request_fire handle args u32 → u64
- session.rs: all session/stream handles u64; uses lookup_session,
  remove_session, session_ep_idx from stream registry
- lib.rs: FfiResponse, FfiDuplexStream, RequestPayload handle fields u32 → u64
- tests/integration.rs: alloc_fetch_token(0) updated call site

Adapters:
- iroh-http-node: BigInt handles in Rust (napi-rs), index.d.ts updated,
  lib.ts bridge methods and session fns updated to bigint
- iroh-http-deno: dispatch.rs structs/params u32 → u64; lib.rs FFI symbol
  u32 → u64; adapter.ts bridge/session/rawFetch updated to bigint
- iroh-http-py: IrohResponse, IrohRequest, IrohBidiStream, IrohUniStream,
  IrohSession handle fields u32 → u64; send_500 signature updated

TypeScript (iroh-http-shared):
- bridge.ts: Bridge interface, FfiResponse, RequestPayload, RawFetchFn,
  AllocBodyWriterFn, FfiDuplexStream all use bigint for handles
- streams.ts: makeReadable/pipeToWriter handle: bigint
- fetch.ts: reqBodyHandle: bigint | null
- serve.ts: resTrailersHandle sentinel 0 → 0n
- session.ts: RawSessionFns interface and buildSession sessionHandle: bigint
- adapter.ts (deno): FFI symbol u64, bridge/session/rawFetch bigint

+++ packages/iroh-http-py/src/lib.rs
@@ -33,7 +33,7 @@ fn py_err(e: impl std::fmt::Display) -> PyErr {
 struct IrohResponse {
     status: u16,
     headers: Vec<(String, String)>,
-    body_handle: u32,
+    body_handle: u64,
     url: String,
 }
 
@@ -167,7 +167,7 @@ impl HandlerResponse {
 /// Incoming request passed to the `serve` handler.
 #[pyclass]
 struct IrohRequest {
-    pub req_body_handle: u32,
+    pub req_body_handle: u64,
     pub method: String,
     pub url: String,
     pub headers: Vec<(String, String)>,
@@ -220,8 +220,8 @@ impl IrohRequest {
 /// and `close()` when done.
 #[pyclass]
 struct IrohBidiStream {
-    read_handle: u32,
-    write_handle: u32,
+    read_handle: u64,
+    write_handle: u64,
 }
 
 // ── IrohUniStream ───────────────────────────────────────────────────────────
@@ -231,7 +231,7 @@ struct IrohBidiStream {
 /// Use `write(data)` to send and `close()` when done.
 #[pyclass]
 struct IrohUniStream {
-    write_handle: u32,
+    write_handle: u64,
 }
 
 #[pymethods]
@@ -344,7 +344,7 @@ impl IrohBrowseSession {
 /// Use `create_bidirectional_stream()` to open streams.
 #[pyclass]
 struct IrohSession {
-    session_handle: u32,
+    session_handle: u64,
 }
 
 #[pymethods]
@@ -878,7 +878,7 @@ async fn handle_request(handler: Arc<PyObject>, payload: iroh_http_core::Request
     }
 }
 
-fn send_500(req_handle: u32, res_body_handle: u32) {
+fn send_500(req_handle: u64, res_body_handle: u64) {
     let _ = respond(req_handle, 500, vec![]);
     let _ = finish_body(res_body_handle);
 }
